refactor: 상행역과 하행역이 같은지 검증을 도메인에서 하도록 수정
