+++ src/main/java/wooteco/subway/infrastructure/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.domain;
+package wooteco.subway.infrastructure;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
