+++ src/main/java/wooteco/subway/line/domain/Section.java
@@ -1,5 +1,6 @@
 package wooteco.subway.line.domain;
 
+import wooteco.subway.line.exception.InvalidSectionRequestException;
 import wooteco.subway.station.domain.Station;
 
 public class Section {
@@ -20,9 +21,16 @@ public class Section {
     }
 
     public Section(Station upStation, Station downStation, int distance) {
+        validate(upStation, downStation);
         this.upStation = upStation;
         this.downStation = downStation;
         this.distance = distance;
+    }
+
+    private void validate(Station upStation, Station downStation) {
+        if (upStation.hasEqualIdWith(downStation)) {
+            throw new InvalidSectionRequestException("상행역과 하행역은 같을 수 없습니다.");
+        }
     }
 
     public Long getId() {
