+++ src/main/java/wooteco/subway/path/ui/PathController.java
@@ -5,13 +5,10 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.auth.domain.AuthenticationPrincipal;
+import wooteco.subway.infrastructure.AuthenticationPrincipal;
 import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.path.application.PathService;
 import wooteco.subway.path.dto.PathResponse;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
